requirements.txt
trimesh
numpy
Pillow